Update text inputs to indicate all orange and tomato are sold out and will be available in winter 2024; replace image file paths with new ones in the HTML file
This commit updates the text inputs for orange and tomato to indicate that they are sold out and will be available again in winter 2024. It also replaces the image file paths for IMG\_9129.PNG and IMG\_9130.PNG with new ones in the HTML file. The old image files have been deleted.